Move the exception to the correct package.
